Allow for floating-point mastery scores
Allow for non-whole number mastery cutoff for teachers. This applies
to _all_ courses in their account.

Fix bug where cancelling the preference edit form returned
an empty response.

+++ app/schemas.py
@@ -4,7 +4,7 @@ from marshmallow import fields, Schema
 class UserPrefsSchema(Schema):
     id = fields.Int(dump_only=True)
     score_calculation_method = fields.Str()
-    mastery_score = fields.Int()
+    mastery_score = fields.Float()
 
 
 class UserLoginSchema(Schema):
